refactor(mbk): migrate auth-store + StepUp surface to @platform/ui
MBK was running byte-identical forks of the shared auth-store + the
entire StepUp surface (modal, controller, interceptor, types, errors).
The forks existed as placeholders waiting for the React 18 → 19
migration, which completed last week (PR #589). Now MBK consumes the
shared @platform/ui implementations and the forks are deleted.

Changes:

1. packages/shared-frontend/src/auth/StepUpModal.tsx
   Apply the React 19 lint fix from PR #604 (key-based remount of the
   StepUpForm child to replace setState-in-effect). MBK had this fix on
   its fork; the shared version did not. Now both have it.

2. apps/mybookkeeper/frontend/src/App.tsx
   Import StepUpModal + useIsAuthenticated from @platform/ui.

3. apps/mybookkeeper/frontend/src/shared/lib/api.ts
   Import installStepUpInterceptor + notifyAuthChange from @platform/ui.

4. apps/mybookkeeper/frontend/src/shared/lib/auth.ts
   Import notifyAuthChange + useIsAuthenticated from @platform/ui;
   keep the existing re-exports so downstream callers don't change.

5. Test updates: vi.mock targets switched from "@/shared/lib/auth-store"
   to "@platform/ui" with importOriginal() preserving the rest of the
   shared package's exports.

6. Deletes (10 files):
   - apps/mybookkeeper/frontend/src/shared/auth/StepUpModal.tsx
   - apps/mybookkeeper/frontend/src/shared/auth/stepUpController.ts
   - apps/mybookkeeper/frontend/src/shared/auth/stepUpInterceptor.ts
   - apps/mybookkeeper/frontend/src/shared/auth/types/StepUpControllerState.ts
   - apps/mybookkeeper/frontend/src/shared/auth/errors/StepUpCancelledError.ts
   - apps/mybookkeeper/frontend/src/shared/auth/errors/StepUpReauthRequiredError.ts
   - apps/mybookkeeper/frontend/src/shared/lib/auth-store.ts
   - apps/mybookkeeper/frontend/src/__tests__/auth-store.test.ts
   - empty parent directories (types/, errors/, auth/)

7. apps/mybookkeeper/scripts/test-map.json: drop deleted-file references.

Validation:
- npm run build (MBK): clean
- npm run build (MJH): clean (consumes shared StepUpModal — confirmed
  no regression from the refactor)
- npm run lint (MBK): exit 0, 0 errors (8 pre-existing framework-level
  warnings unchanged)
- auth.test.ts + apiInterceptor.test.ts + InviteAccept.test.tsx
  (22 tests): all passing
- shared-frontend tests: 143 passing (2 pre-existing failures in
  InlineBoldText, unrelated)
- Full MBK suite: 1748 passing (9 pre-existing failures in unrelated
  test files — Documents, hourly-rate, LeaseTemplateUploadDialog,
  PromoteFromInquiryPanel, PublicInquiryForm — verified pre-existing
  on main during PR #595 work)

Closes the parity gap noted in the MBK CLAUDE.md fork comment and in
project_mbk_platform_ui_migration_blocked.md (which was blocked on
React 19; that block lifted with PR #589).
