+++ apps/mybookkeeper/frontend/src/App.tsx
@@ -2,14 +2,13 @@ import { lazy, Suspense } from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "@/shared/store";
-import { useIsAuthenticated } from "@/shared/lib/auth";
+import { StepUpModal, useIsAuthenticated } from "@platform/ui";
 import Layout from "@/app/components/Layout";
 import RequireRole from "@/shared/components/RequireRole";
 import RequireOrg from "@/shared/components/RequireOrg";
 import RequireOrgRole from "@/shared/components/RequireOrgRole";
 import ErrorBoundary from "@/shared/components/ErrorBoundary";
 import Toaster from "@/shared/components/ui/Toaster";
-import StepUpModal from "@/shared/auth/StepUpModal";
 import Login from "@/app/pages/Login";
 import Register from "@/app/pages/Register";
 import ForgotPassword from "@/app/pages/ForgotPassword";
